Move Sentry Under Analytics
