+++ src/aquarius.js
@@ -4,6 +4,7 @@ import Discord from 'discord.js';
 import fs from 'fs';
 import yaml from 'js-yaml';
 import path from 'path';
+import Sentry from './lib/analytics/sentry';
 import Analytics from './lib/commands/analytics';
 import Settings from './lib/commands/settings';
 import * as loading from './lib/core/loading';
@@ -11,7 +12,6 @@ import * as permissions from './lib/core/permissions';
 import * as triggers from './lib/core/triggers';
 import database from './lib/database/database';
 import { fixPartialReactionEvents } from './lib/discord/library-fixes';
-import Sentry from './lib/errors/sentry';
 import { isBot, isDirectMessage } from './lib/helpers/messages';
 import DirectMessageManager from './lib/managers/direct-message-manager';
 import EmojiManager from './lib/managers/emoji-manager';
