+++ src/global/commands/admin.js
@@ -1,7 +1,7 @@
 import debug from 'debug';
 import dedent from 'dedent-js';
 import pluralize from 'pluralize';
-import Sentry from '../../lib/errors/sentry';
+import Sentry from '../../lib/analytics/sentry';
 import { humanize } from '../../lib/helpers/lists';
 
 const log = debug('Admin');
