+++ src/lib/helpers/embeds.js
@@ -1,7 +1,7 @@
 import debug from 'debug';
 import dedent from 'dedent-js';
 import { RichEmbed } from 'discord.js';
-import Sentry from '../errors/sentry';
+import Sentry from '../analytics/sentry';
 import { getIconColor } from './colors';
 import { getStandardDate } from './dates';
